parallel migration
Former-commit-id: 847778cb720cb5427998c6097181026671f6ef54

+++ api/c/parallel.c
@@ -81,7 +81,7 @@ char** sf_split(sf_file inp          /* input file */,
     
     jobs = nodes-1;
     if (jobs < split) {
-	w = 1+(float) split/jobs;
+	w = (int) (1+((float) split)/jobs);
     } else {
 	w = 1;
 	jobs = split;
